APK Extraction

+++ Ornate.Lite/MainWindow.xaml.cs
@@ -4,12 +4,15 @@ using Avalonia.Dialogs;
 using Avalonia.Interactivity;
 using Avalonia.Markup.Xaml;
 using Avalonia.Threading;
+using Ornate.Lite.Dialogs;
 using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using Xilium.CefGlue.Common;
 
 namespace Ornate.Lite
@@ -25,7 +28,6 @@ namespace Ornate.Lite
         {
             AvaloniaXamlLoader.Load(this);
 
-            return;
             CreateBrowserView();
 
             var mainMenu = this.FindControl<Menu>("mainMenu");
@@ -87,16 +89,36 @@ namespace Ornate.Lite
 
             if (Directory.Exists(BrowserView.DataPath))
             {
-                //TODO: warn user + let them confirm
-                ConfirmationWindow dialog = new()
+                // warn user + let them confirm
+                ConfirmationWindow warningDialog = new()
                 {
                     Title = "Warning!",
                     Prompt = "You are about to delete and replace your existing content. Are you sure?"
                 };
-                await dialog.ShowDialog(this);
-                return;
-                // delete folder
-                Directory.Delete(BrowserView.DataPath, true);
+                var result = await warningDialog.ShowDialog<bool?>(this);
+                // user canceled or aborted
+                if (!result.HasValue || !result.Value)
+                    return;
+
+                // Inform the user that we're deleting files, so it doesnt look like the thing just lags
+                InfoWindow deletingFilesDialog = new()
+                {
+                    Title = "Deleting...",
+                    Prompt = "Deleting old files..."
+                };
+                // Run the deletion inside a new thread to not block the UI thread
+                deletingFilesDialog.Opened += (_, _) =>
+                {
+                    var worker = new Thread(() =>
+                    {
+                        // Delete the folder
+                        Directory.Delete(BrowserView.DataPath, true);
+                        // Close the dialog from the UI thread
+                        Dispatcher.UIThread.Post(() => deletingFilesDialog.Close());
+                    });
+                    worker.Start();
+                };
+                await deletingFilesDialog.ShowDialog(this);
             }
 
             //TODO: kill browser beforehand?
@@ -106,32 +128,70 @@ namespace Ornate.Lite
             // find all entries in "assets" folder and below
             var assets = zip.Entries.Where(e => e.FullName.StartsWith(assetsFolder));
             if (!assets.Any()) // didnt find anything
-                return; //TODO: warning
-
-            //TODO: progress bar
-            foreach (var asset in assets)
             {
-                // put the path together: "data/" + path of file we're looking at (removing the "assets/" at the start)
-                var path = Path.Combine(BrowserView.DataPath, asset.FullName.Substring(assetsFolder.Length));
-                // check if parent dir exists => if not create it
-                if (!Directory.Exists(path))
-                    Directory.CreateDirectory(Path.GetDirectoryName(path));
-
-                asset.ExtractToFile(path, true);
+                // inform the user about this
+                OKWindow noAssetsFoundWindow = new()
+                {
+                    Title = "Error",
+                    Prompt = $"Didn't find a 'assets' folder inside {files[0]}."
+                };
+                await noAssetsFoundWindow.ShowDialog(this);
+                return;
             }
-            //TODO: done info
+
+            // create a progress bar for the user
+            ProgressBarWindow progressBar = new()
+            {
+                Title = "Extracting files...",
+                Text = "Extracting files...",
+                Min = 0,
+                Max = 100,
+                Value = 0,
+            };
+            progressBar.Opened += (_, _) =>
+            {
+                // Run the extraction in a new thread to not block the UI thread
+                var worker = new Thread(() =>
+                {
+                    var total = (float)assets.Count();
+                    var current = 0f;
+                    foreach (var asset in assets)
+                    {
+                        // put the path together: "data/" + path of file we're looking at (removing the "assets/" at the start)
+                        var path = Path.Combine(BrowserView.DataPath, asset.FullName.Substring(assetsFolder.Length));
+                        // check if parent dir exists => if not create it
+                        if (!Directory.Exists(path))
+                            Directory.CreateDirectory(Path.GetDirectoryName(path));
+
+                        asset.ExtractToFile(path, true);
+                        current++;
+                        // Update the progress bar in the ui thread
+                        Dispatcher.UIThread.Post(() => progressBar.Value = (int)((current / total)*100));
+                    }
+                    // Close the dialog from the UI thread
+                    Dispatcher.UIThread.Post(() => progressBar.Close());
+                });
+                worker.Start();
+            };
+            await progressBar.ShowDialog(this);
+
+            // show done info
+            OKWindow okWindow = new()
+            {
+                Title = "Done",
+                Prompt = $"Successfully extracted all files to {Path.GetFullPath(BrowserView.DataPath)}"
+            };
+            await okWindow.ShowDialog(this);
         }
 
         private async void OnDebugMenuItemClick(object sender, RoutedEventArgs e)
         {
             //TODO: remove when done
-            ConfirmationWindow dialog = new()
-            {
-                Title = "Warning!",
-                Prompt = "You are about to delete and replace your existing content. Are you sure?"
-            };
-            await dialog.ShowDialog(this);
-            //TODO: result
+        }
+
+        private void Dialog_Opened(object sender, EventArgs e)
+        {
+            throw new NotImplementedException();
         }
 
         private void OnReloadGameMenuItemClick(object sender, RoutedEventArgs e) => OnReloadGameNativeMenuItemClick(sender, e);
